# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend normativo de acceso a derechos",
    "description": "\nAPI de consulta del corpus normativo de acceso a derechos.\n\nToda respuesta lleva la misma envoltura: `schema_version`, `release_id`,\n`as_of`, `known_at`, `data_status`, `evidence`, `missing_fields` y `warnings`.\nSin `as_of` una respuesta no dice para cuándo vale; sin `release_id` no se puede\nreproducir; sin `missing_fields` una abstención sería indistinguible de un \"no\".\n\nLos errores son tipados. Información insuficiente es un estado de dominio y\nnunca se disfraza de 500.\n\nEsta API no otorga, no rechaza y no revoca prestaciones. Devuelve lo que dicen\nlas normas publicadas, con la evidencia que lo sostiene y lo que falta averiguar.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/recuperacion": {
      "post": {
        "tags": [
          "recuperación"
        ],
        "summary": "Recuperar",
        "operationId": "recuperar_v1_recuperacion_post",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha para la que se consulta. Por omisión, hoy.",
              "title": "As Of"
            },
            "description": "Fecha para la que se consulta. Por omisión, hoy."
          },
          {
            "name": "known_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada.",
              "title": "Known At"
            },
            "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudRecuperacion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_ResultadoRecuperacion_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cobertura": {
      "get": {
        "tags": [
          "recuperación"
        ],
        "summary": "Cobertura",
        "description": "Denominadores y pendientes, separados de lo validado y publicable.",
        "operationId": "cobertura_v1_cobertura_get",
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha para la que se consulta. Por omisión, hoy.",
              "title": "As Of"
            },
            "description": "Fecha para la que se consulta. Por omisión, hoy."
          },
          {
            "name": "known_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada.",
              "title": "Known At"
            },
            "description": "Instante de conocimiento: qué sabía el sistema en ese momento. Permite reproducir una respuesta pasada."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Respuesta_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Advertencia": {
        "properties": {
          "codigo": {
            "$ref": "#/components/schemas/CodigoError"
          },
          "detalle": {
            "type": "string",
            "title": "Detalle"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "detalle"
        ],
        "title": "Advertencia"
      },
...
      "CodigoError": {
        "type": "string",
        "enum": [
          "SOURCE_UNAVAILABLE",
          "STALE_DATA",
          "CONFLICT",
          "INSUFFICIENT_EVIDENCE",
          "UNKNOWN_IDENTITY",
          "UNSUPPORTED_SCOPE",
          "INVALID_REQUEST",
          "NOT_AUTHORIZED",
          "VERSION_CONFLICT"
        ],
        "title": "CodigoError",
        "description": "Errores tipados del contrato.\n\nInformación insuficiente es un estado de dominio, no un fallo del servidor:\nnunca se disfraza de 500."
      },
...
      "DataStatus": {
        "type": "string",
        "enum": [
          "PUBLICADO",
          "PARCIAL",
          "NO_PUBLICABLE",
          "SIN_RESULTADOS"
        ],
        "title": "DataStatus",
        "description": "Qué tan sostenible es lo que devuelve la respuesta."
      },
      "Evidencia": {
        "properties": {
          "evidencia_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidencia Id"
          },
          "fragmento": {
            "type": "string",
            "title": "Fragmento"
          },
          "norma": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Norma"
          },
          "unidad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidad"
          },
          "url_fuente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Fuente"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "capturado_en": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capturado En"
          }
        },
        "type": "object",
        "required": [
          "evidencia_id",
          "fragmento"
        ],
        "title": "Evidencia",
        "description": "Localizador verificable de una afirmación."
      },
      "Fragmento": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chunk Id"
          },
          "texto": {
            "type": "string",
            "title": "Texto"
          },
          "norma": {
            "type": "string",
            "title": "Norma"
          },
          "unidad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidad"
          },
          "url_fuente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Fuente"
          },
          "relevancia": {
            "type": "number",
            "title": "Relevancia"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "texto",
          "norma",
          "unidad",
          "url_fuente",
          "relevancia"
        ],
        "title": "Fragmento"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Respuesta_ResultadoRecuperacion_": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "release_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Id"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "known_at": {
            "type": "string",
            "format": "date-time",
            "title": "Known At"
          },
          "data_status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "data": {
            "$ref": "#/components/schemas/ResultadoRecuperacion"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidencia"
            },
            "type": "array",
            "title": "Evidence"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Advertencia"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "known_at",
          "data_status",
          "data"
        ],
        "title": "Respuesta[ResultadoRecuperacion]"
      },
      "Respuesta_dict_": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "release_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Id"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "known_at": {
            "type": "string",
            "format": "date-time",
            "title": "Known At"
          },
          "data_status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidencia"
            },
            "type": "array",
            "title": "Evidence"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Advertencia"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "known_at",
          "data_status",
          "data"
        ],
        "title": "Respuesta[dict]"
      },
...
      "ResultadoRecuperacion": {
        "properties": {
          "fragmentos": {
            "items": {
              "$ref": "#/components/schemas/Fragmento"
            },
            "type": "array",
            "title": "Fragmentos"
          },
          "conflictos_pertinentes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Conflictos Pertinentes"
          },
          "aclaracion": {
            "type": "string",
            "title": "Aclaracion",
            "default": "Los fragmentos sirven para explicar y citar. Los montos, las fechas y los datos de contacto se consultan con las operaciones tipadas, no combinando fragmentos."
          }
        },
        "type": "object",
        "title": "ResultadoRecuperacion"
      },
...
      "SolicitudRecuperacion": {
        "properties": {
          "consulta": {
            "type": "string",
            "title": "Consulta"
          },
          "jurisdiccion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiccion"
          },
          "limite": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limite",
            "default": 10
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "consulta"
        ],
        "title": "SolicitudRecuperacion"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}